Move TestNewBindingInfo to its own file
