+++ pkg/controller/servicebindingrequest/olm_test.go
@@ -18,6 +18,11 @@ import (
 	"github.com/redhat-developer/service-binding-operator/test/mocks"
 )
 
+const (
+	dataUserAnnotation     = "servicebindingoperator.redhat.io/status.dbCredentials-data.user"
+	dataPasswordAnnotation = "servicebindingoperator.redhat.io/status.dbCredentials-data.password"
+)
+
 func init() {
 	logf.SetLogger(logf.ZapLogger(true))
 }
@@ -80,57 +85,6 @@ func TestOLMNew(t *testing.T) {
 		assertGVKs(t, gvks)
 	})
 }
-
-// TestNewBindingInfo exercises annotation binding information parsing.
-func TestNewBindingInfo(t *testing.T) {
-	type args struct {
-		s string
-		d string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-		want    *BindingInfo
-	}{
-		{
-			args: args{s: "status.configMapRef-password", d: "binding"},
-			want: &BindingInfo{
-				FieldPath:  "status.configMapRef",
-				Descriptor: "binding:password",
-				Path:       "password",
-			},
-			name:    "{fieldPath}-{path} annotation",
-			wantErr: false,
-		},
-		{
-			args: args{s: "status.connectionString", d: "binding"},
-			want: &BindingInfo{
-				Descriptor: "binding:status.connectionString",
-				FieldPath:  "status.connectionString",
-				Path:       "status.connectionString",
-			},
-			name:    "{path} annotation",
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			b, err := NewBindingInfo(tt.args.s, tt.args.d)
-			if err != nil && !tt.wantErr {
-				t.Errorf("NewBindingInfo() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			} else if err == nil {
-				requireYamlEqual(t, tt.want, b)
-			}
-		})
-	}
-}
-
-var (
-	dataUserAnnotation     = "servicebindingoperator.redhat.io/status.dbCredentials-data.user"
-	dataPasswordAnnotation = "servicebindingoperator.redhat.io/status.dbCredentials-data.password"
-)
 
 func TestOLMBuildDescriptorsFromAnnotations(t *testing.T) {
 	type args struct {
